add aportgen test case again
This test case was removed earlier, due to its strong dependency on the
aports folder. This new version uses a known input and output instead of
working with the real pmaports and upstream aports from Alpine. That way
we can really test if the pmbootstrap code is working, no matter if the
aports changed.