install apk keys and binfmt data with setup.py
So far we've only installed the Python source code in the pmb folder.
But pmbootstrap needs the apk package signing keys as well as the
binfmt data file to work correctly, so this commit moves all of them
to `pmb/data` and changes `setup.py` and `MANIFEST.in` to install them
as well. Also the code and test case have been adjusted to use the new
paths.