pmb.parse: Preferentially use a provider whose origin is installed
If there are multiple packages that provide the same name (eg mesa-gl
and mesa-pvr-gl), pmbootstrap will always pick mesa-gl, even if we're
installing mesa-pvr instead of plain mesa.  Packages have the idea
of an "origin", which is the main package that they're built from.
We can use this to our benefit by picking which provider to install
if the origin package of one is already due to be installed.  If
none of the origin's are set to be installed, then we will continue
to fall back to either the package name directly or the shortest
package name if all else fails.