pmb.config: Add default user to the dialout group
This allows access to serial devices. My reason for wanting this is
because Plasma Bigscreen uses HDMI CEC to allow remote control via the
TV remote, and the CEC adapter on devices appear as /dev/ttyACM0, which
is owned by the dialout group. Without adding the user to this group,
the user can not control their device through HDMI CEC