Add "pmbootstrap check" to run CI scripts locally
fixes #2169

Some parts of this code is kind of hacky but it could be a good start and it
works.

The original issue said "utilizing Alpine chroots as it makes sense".
I should have asked what was meant by this but I moved everything to the chroot
so that our tests can be ran on any distro.

I have used "pmbootstrap check" locally and it's convenient to be able to
run the CI on my fast machine rather than wait for gitlab.

Hope this helps!